refactor: separate trait Cluster and its impl
